Switch to `prepare` npm script
Per the output of `npm publish . --tag latest --dry-run`:

> ERROR npm WARN prepublish-on-install As of npm@5, `prepublish` scripts are deprecated.
> npm WARN prepublish-on-install Use `prepare` for build steps and `prepublishOnly` for upload-only.
> npm WARN prepublish-on-install See the deprecation note in `npm help scripts` for more information.

`prepublishOnly` is the standard replacement for this, and matches what
e.g. `ember-cli-typescript` does, but the build pipeline here is a bit
different from that flow: it generates both CommonJS output and Ember
addon output via a build command. Accordingly `prepare` is the right
option.